requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
boto3
python-dotenv
httpx
openai
google-generativeai
